backend/requirements.txt
fastapi
uvicorn
numpy
pillow
tensorflow-macos
tensorflow-metal
python-multipart
